pyproject.toml
[project]
name = "fish-audio-sdk"
version = "1.0.0"
description = "Python SDK for Fish Audio text-to-speech and voice cloning API"
authors = [
    {name = "abersheeran", email = "me@abersheeran.com"},
    {name = "James Ding", email = "james@fish.audio"},
]
dependencies = [
    "httpx>=0.27.2",
    "ormsgpack>=1.5.0",
    "pydantic>=2.9.1",
    "httpx-ws>=0.6.2",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}
keywords = ["fish-audio", "tts", "text-to-speech", "voice-cloning", "ai", "speech-synthesis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://fish.audio"
Documentation = "https://docs.fish.audio"
Repository = "https://github.com/fishaudio/fish-audio-python"
Issues = "https://github.com/fishaudio/fish-audio-python/issues"

[project.optional-dependencies]
utils = [
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fishaudio", "src/fish_audio_sdk"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.14.3",
]
